Update interop example
