+++ examples/interop/pos-sidechain-example-two/src/app/modules/react/commands/react_cc_command.ts
@@ -12,7 +12,7 @@ import { CROSS_CHAIN_COMMAND_REACT } from '../constants';
 import { CCReactCommandParamsSchema, CCReactMessageParamsSchema } from '../schemas';
 import { CCReactMessageParams, CCReactCommandParams, InteroperabilityMethod } from '../types';
 
-export class ReactCrossChainCommand extends BaseCommand {
+export class CrossChainReactCommand extends BaseCommand {
 	private _interoperabilityMethod!: InteroperabilityMethod;
 	public schema = CCReactCommandParamsSchema;
